% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{absolute_error}
\alias{absolute_error}
\title{Absolute Error of glmnet Models}
\usage{
absolute_error(mdl, x, y)
}
\arguments{
\item{mdl}{a previously fitted glmnet model of any type except "cox"}

\item{x}{a vector of samples for which to make predictions}

\item{y}{the actual response values of the samples in x}
}
\value{
a matrix containing the absolute error for each lambda value in
  the original model. Matrix has one row per sample and one column per lambda.
}
\description{
This function calculates the absolute error a glmnet model on a given
set of samples. This is particularly useful for calculating model performance
on hold-out sets. Note that this is NOT producing the mean absolute error!
}
