% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{binomial_deviance}
\alias{binomial_deviance}
\title{Binomial Deviance Function Stolen from Glmnet}
\usage{
binomial_deviance(mdl, x, y)
}
\arguments{
\item{mdl}{a previously fitted glmnet model}

\item{x}{a vector of values for which to make predictions}

\item{y}{the reference category assignments for the samples in x}
}
\value{
a matrix containing the deviance values for each lambda value in the
  original model. Matrix has one row per sample and one column per lambda.
}
\description{
The glmnet package doesn't give us the individual sample deviance values, so
I've ripped out the relevant function and put it here. This function
calculates the binomial deviance for a glmnet model on a given set of
samples. This is particularly useful for calculating model performance on
hold-out sets.
}
