% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{calc_roc}
\alias{calc_roc}
\title{Calculate and Print ROC Plots}
\usage{
calc_roc(ctrl_group, case_group)
}
\arguments{
\item{ctrl_group}{a vector of values taken by the control samples}

\item{case_group}{a vector of values taken by the case samples}
}
\value{
A list containing the ROC curve points, the ROC AUC, and the Delong
confidence interval for the AUC.
}
\description{
Calculates a ROC plot from case and control group values. Plots the ROC curve
and its sensitivity confidence interval. Returns the ROC curve points and
AUC value.
}
