% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plots.R
\name{calculate_volcano}
\alias{calculate_volcano}
\title{Generate data for a volcano plot}
\usage{
calculate_volcano(
  feature_data,
  sample_labels,
  test = stats::t.test,
  p_adjust = "none"
)
}
\arguments{
\item{feature_data}{Feature values for all samples, where rows are samples
and columns are features.}

\item{sample_labels}{Factor with two levels specifying which group each.sample
belongs to.}

\item{test}{The statistical test function to use for calculating p-values.}

\item{p_adjust}{The type of multiple comparisons correction to perform.}
}
\value{
Tibble with columns 'feature', 'fold_change', 'p_value', and 'p_adj'
}
\description{
Given a dataset with two labels, calculates the fold-change and p-values
for each feature in the dataset.
}
