% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{coef.multi.cv.glmnet}
\alias{coef.multi.cv.glmnet}
\title{Extract Coefficients from a "multi.cv.glmnet" Object}
\usage{
\method{coef}{multi.cv.glmnet}(object, s = "lambda.1se", exact = FALSE, ...)
}
\arguments{
\item{object}{a multi.cv.glmnet object from which to extract coefficients}

\item{s}{the desired lambda value to use. Special values "lambda.min" and
"lambda.1se" automatically select the corresponding values from the
multi.cv.glmnet object.}

\item{exact}{if the desired lambda value was not part of the original
sequence, should the models be refit to exactly calculate the coefficients?
See [glmnet::coef.glmnet()] for details.}

\item{...}{not used}
}
\value{
a tibble containing every non-zero feature coefficient from every
  cross-validation model
}
\description{
This function extracts the model coefficients at a single lambda value from
each of the models in a multi.cv.glmnet object. This is particularly useful
for evaluating model stability and selecting robust features for a final
model. If the models only consistently select a small subset of features, it
is often useful to fit a non-sparse logistic model using just those features.
}
