% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrtools.R
\name{concat_file_paths}
\alias{concat_file_paths}
\title{let's parallelize concat some data.tables}
\usage{
concat_file_paths(filepaths, cores = 1)
}
\arguments{
\item{filepaths}{vector of rds filepaths}

\item{cores}{number of cores to mclapply over lol}
}
\value{
data.table
}
\description{
function to parallelize and concatenate a list of vector file paths that contain rds that hold only data.tables
}
