% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ms_alignment.R
\name{ConvertIntToUInt}
\alias{ConvertIntToUInt}
\title{Fix a 32 bit unsigned integer that has been read as signed}
\usage{
ConvertIntToUInt(x, adjustment = 2^32)
}
\arguments{
\item{x}{Number to be fixed}

\item{adjustment}{number to be added to convert to uint32 (2^32 by default)}
}
\value{
numeric value of uint32
}
\description{
This is really just to fix a limitation of readBin/R's 32 bit signed ints
}
\seealso{
\code{\link{readBin}}
}
\author{
jefferis
}
