% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Calculate Shannon Entropy}
\usage{
entropy(x)
}
\arguments{
\item{x}{The data set in a format acceptable for the table function}
}
\value{
the number of bits of entropy (log base 2) in the data set
}
\description{
Calculates the Shannon entropy of a set of data.
}
