% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gap_fill_status.R
\name{extract_gap_fill_status}
\alias{extract_gap_fill_status}
\title{Extract gap fill status information for compounds in a Compound Discoverer alignment}
\usage{
extract_gap_fill_status(msa, ids = NULL, detailed = FALSE)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{ids}{Compound IDs to get gap status info for, or NULL to return all}

\item{detailed}{Should detailed gap filling status be returned? Defaults to
simple status.}
}
\value{
A tibble containing the gap fill status data
}
\description{
Compound Discoverer mass spec data alignment databases store gap filling data
for all consolidated compounds as binary blobs. This function extracts that
information and converts it into human readable factor data. Note that this
data is only for the reference/best ion in each compound. Information for
other ions must be extracted from the enormous MissingCompoundIonInstanceItems
table.
}
\details{
The alignment offers two levels of status information. The simple one has
levels "No gap", "Missing ions", and "Full gap". The more detailed status
includes information on exactly how any gaps were filled. In the detailed
status information, 'unknown status' seems to correspond to features that
did not require gap filling.
}
\note{
Compound Discoverer software is produced by Thermo Fisher Scientific. This
package is not affiliated with Thermo Fisher Scientific in any way. For an
official Python interface to Compound Discoverer alignment files, see
\url{https://github.com/thermofisherlsms/pyeds}
}
