% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_alignment_functions.R
\name{extract_isotope_pattern_blob}
\alias{extract_isotope_pattern_blob}
\title{Extract predicted isotopologue pattern data from an SQLite blob}
\usage{
extract_isotope_pattern_blob(blb, zip_dir = tempdir())
}
\arguments{
\item{blb}{A raw vector containing an SQLite blob of isotopologue data}

\item{zip_dir}{Directory where XML files should be unzipped}
}
\value{
A tibble containing the isotopologue pattern
}
\description{
Certain mass spec data alignment formats store predicted isotopologue spectra
as SQLite blobs containing a compressed XML file. This function extracts the
stored data from such blobs.
}
