% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_peak_model_blob.R
\name{extract_peak_model_blob}
\alias{extract_peak_model_blob}
\title{Extract fitted peak model parameters from Compound Discoverer alignment}
\usage{
extract_peak_model_blob(blb, zip_dir = tempdir())
}
\arguments{
\item{blb}{A raw vector containing an SQLite blob of a peak model}

\item{zip_dir}{Directory where XML files should be unzipped}
}
\value{
The XML structure describing the peak model.
}
\description{
Compound Discoverer cdResult alignment databases store fitted peak models as
SQLite blobs containing a compressed XML file. This function extracts the
stored data from such blobs.
}
\details{
Note: specifying a temp directory will speed up extraction of multiple blobs.
}
\note{
Compound Discoverer software is produced by Thermo Fisher Scientific. This
package is not affiliated with Thermo Fisher Scientific in any way. For an
official Python interface to Compound Discoverer alignment files, see
\url{https://github.com/thermofisherlsms/pyeds}
}
