% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_peak_ratings.R
\name{extract_peak_ratings}
\alias{extract_peak_ratings}
\title{Extract peak quality blobs in a Compound Discoverer alignment}
\usage{
extract_peak_ratings(msa, ids = NULL)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{ids}{Compound IDs to get ratings for, or NULL to get all peak ratings}
}
\value{
A tibble containing the peak ratings
}
\description{
Compound Discoverer cdResult alignment databases store peak quality ratings
as SQLite blobs containing the data in binary form. This function extracts the
stored data from such blobs. The returned ratings are for the reference adduct
ion for each compound. Other areas must be retrieved manually from the relevant
tables.
}
\note{
Compound Discoverer software is produced by Thermo Fisher Scientific. This
package is not affiliated with Thermo Fisher Scientific in any way. For an
official Python interface to Compound Discoverer alignment files, see
\url{https://github.com/thermofisherlsms/pyeds}
}
