% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_proximity_oob.R
\name{extract_proximity_oob}
\alias{extract_proximity_oob}
\title{Calculate OOB proximity from ranger random forest model}
\usage{
extract_proximity_oob(rf_fit, old_data)
}
\arguments{
\item{rf_fit}{a ranger model}

\item{old_data}{The data used to fit the original model}
}
\value{
A symmetric matrix containing the proximity values.
}
\description{
Given an existing random forest model generated by the ranger package,
calculate the proximity between each pair of features in the original model.
This is only calculated using trees where _both_ samples are out-of-bag to
avoid overly optimistic proximity measures.
}
\details{
Code adapted from \url{https://github.com/imbs-hl/ranger/issues/234}
}
