% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_xic_blob.R
\name{extract_xic_blob}
\alias{extract_xic_blob}
\title{Extract XIC data from an SQLite blob}
\usage{
extract_xic_blob(blb)
}
\arguments{
\item{blb}{A raw vector containing an SQLite blob of XIC data}
}
\value{
A tibble containing the data
}
\description{
Certain mass spec data alignment formats store feature extracted ion
chromatograms as SQLite blobs containing a compressed binary file.
This function extracts the stored data from such blobs.
}
\details{
The precise format of the binary file has not yet been completely elucidated.
Chunk 1 looks like run time. Chunk 2 is unknown, maybe resolution? Chunk 3
is obviously the XIC itself. Chunk 4 looks like probably the instantaneous
noise level, but might be something else. Subtracting chunk 4 from chunk 3
seems to yield reasonable looking XIC traces.
}
