% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{fancy_roc_plot}
\alias{fancy_roc_plot}
\title{Construct fancy-looking ROC plot}
\usage{
fancy_roc_plot(data)
}
\arguments{
\item{data}{output from the [calc_roc] function}
}
\value{
a ggplot object containing the plot, so you can fine-tune it as
  desired before displaying it
}
\description{
Generates a ggplot object to display a ROC curve.
}
\details{
There isn't really any error-checking here. Put garbage in, you'll
  get garbage out.
}
