% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{find_kplexes}
\alias{find_kplexes}
\title{Find connected k-plexes in networks}
\usage{
find_kplexes(g, min_kc, kp)
}
\arguments{
\item{g}{the igraph object to search for k-plexes}

\item{min_kc}{The minimum coreness to consider}

\item{kp}{the maximum number of missing edges per node in the plex}
}
\value{
a list containing a list of the k-plex graphs, the minimum degrees of
  each plex, and the sizes of each plex
}
\description{
Given an igraph object, this function identifies connected components of
induced subgraphs where all N nodes in the component are connected to at
least N-k other nodes. Thus, a 1-plex is the same as a clique of size N,
because all nodes are connected to each other but not to themselves. A 2-plex
is a clique where some nodes are missing at most one edge to another node.
}
\details{
WARNING: this method uses an untested algorithm! Use at your own risk! Only
works for connected k-plexes! May not find maximal k-plexes or all k-plexes.
Requires more formalized testing.

This is a really simplistic kludge. Use at your own risk! May not
  behave as intended!
}
