% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sim_pairs.R
\name{find_sim_pairs}
\alias{find_sim_pairs}
\title{Memory-efficient search for similar numeric values}
\usage{
find_sim_pairs(x, maxDiff, ppm = FALSE)
}
\arguments{
\item{x}{sorted numeric vector to be grouped}

\item{maxDiff}{maximum difference between pairs}

\item{ppm}{FALSE (default) to use absolute differences, or TRUE to use
parts-per-million differences instead}
}
\value{
A two-column matrix giving the indices of all pairs.
}
\description{
Given a sorted numeric vector, this function finds all pairs of values that
are within a threshold difference of each other.
}
