% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_source_decay_families.R
\name{find_source_decay_families}
\alias{find_source_decay_families}
\title{Find putative source decay families in mass spec alignments}
\usage{
find_source_decay_families(
  msa,
  raw_data_folder,
  ref_align_file,
  drt_max = 0.02,
  area_cor_min = 0.997,
  xic_tol = 5,
  xic_error_max = 0.3,
  bp_param = BiocParallel::SerialParam()
)
}
\arguments{
\item{msa}{An ms_alignment object containing the data to be analyzed.}

\item{raw_data_folder}{A folder where the raw MS data files are stored in mzML format.}

\item{ref_align_file}{The base file name (no path, no extension) of the MS data file to use as reference for retention time alignment.}

\item{drt_max}{Maximum retention time difference (in minutes) to consider.}

\item{area_cor_min}{Minimum peak area correlation to consider.}

\item{xic_tol}{Mass tolerance for extracted ion chromatogram calculation (XIC) in parts per million (ppm).}

\item{xic_error_max}{Maximum XIC error to consider (range 0-1).}

\item{bp_param}{Optional BiocParallelParam object describing parallel processing to be used. Defaults to serial processing if not supplied.}
}
\value{
A list containing:\tabular{ll}{
   \code{matched_compound_pairs} \tab Table of all matching pairs of compounds. \cr
   \code{sdf_graph} \tab Network object constructed from compound pairs. \cr
   \code{sdf_cluster_members} \tab Table of cluster member IDs and associated cluster IDs. \cr
   \code{sdf_cluster_centers} \tab Compound IDs of the central member of each compound cluster. \cr
   \tab \cr
}
}
\description{
Some compounds fragment when ionized into the gas phase, forming a source
decay family. Mass spec features in these families are essentially duplicates
of the parent compound and should be identified and treated specially. This
function identifies such families based on very similar retention times,
correlated peak areas across all samples, and identical peak shapes. This
approach also finds instances where the alignment software has missed adducts,
isotopologues, or multiply charged versions of a single compound. It also
corrects for occasional alignment software errors.
}
