% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_entropy_search.R
\name{flash_entropy_search}
\alias{flash_entropy_search}
\title{Perform flash entropy search}
\usage{
flash_entropy_search(fragment_library, query_spectrum, ms2_tol_ppm = 5)
}
\arguments{
\item{fragment_library}{matrix of library spectra sorted by fragment m/z. Must
have columns "Spectrum", "mz", and "intensity"}

\item{query_spectrum}{matrix with query spectrum. Must have columns "mz" and
"intensity"}

\item{ms2_tol_ppm}{Fragment matching tolerance in PPM}
}
\value{
A tibble containing spectral entropy similarity scores between the
  query spectrum and any library spectra with matching peaks.
}
\description{
Calculates flash entropy similarity between a query spectrum and a library of
reference spectra. By converting the library and query m/z values to neutral
losses, one can also do neutral loss matching.
}
\details{
Current implementation only searches one spectrum in 'open'/similarity mode.
}
