% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_fragment_search.R
\name{flash_fragment_search}
\alias{flash_fragment_search}
\title{Quickly search spectral library for fragment pattern}
\usage{
flash_fragment_search(fragment_library, query_ions, ms2_tol_ppm = 5)
}
\arguments{
\item{fragment_library}{matrix of compound spectra sorted by fragment m/z.
Must have columns "Spectrum", "mz", and "intensity"}

\item{query_ions}{matrix with query fragment ions. Must have columns "mz" and
"intensity"}

\item{ms2_tol_ppm}{Fragment matching tolerance in PPM}
}
\value{
A tibble containing the number of query ions found in each compound
  spectrum.
}
\description{
Given a library of compound spectra and a set of fragment ions, this function
identifies all compounds containing one or more of the fragment ions. By
converting the library and fragment ion m/z values to neutral losses, one can
also do neutral loss searching
}
\details{
Current implementation does not consider the relative fragment intensities.
}
