% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_alignment_functions.R
\name{get_all_spectra}
\alias{get_all_spectra}
\title{Retrieve all spectra in alignment}
\usage{
get_all_spectra(msa)
}
\arguments{
\item{msa}{An ms_alignment object to query}
}
\value{
A tibble with the spectra
}
\description{
Queries a mass spec alignment database to get all the mass spectra in it. The
actual spectra are stored as XML blobs which can be extracted using
extract_spectral_blob.
}
