% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdresult_enum.R
\name{get_cdresult_enum}
\alias{get_cdresult_enum}
\title{Retrieve factor labels in Compound Discoverer alignment results}
\usage{
get_cdresult_enum(msa, table, enum)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{table}{Name of data table where enum is located}

\item{enum}{Name of the enum/factor column}
}
\value{
A tibble containing information on each value of the enum. If
  the column is not an enum, returns NULL with a warning.
}
\description{
CompoundDiscoverer alignment databases contain many enum/factor columns
encoded as integer values. This function looks up the labels for a given
factor in a particular table.
}
\note{
Compound Discoverer software is produced by Thermo Fisher Scientific. This
package is not affiliated with Thermo Fisher Scientific in any way. For an
official Python interface to Compound Discoverer alignment files, see
\url{https://github.com/thermofisherlsms/pyeds}
}
