% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdresult_tables.R
\name{get_cdresult_tables}
\alias{get_cdresult_tables}
\title{Retrieve related tables in Compound Discoverer alignment results}
\usage{
get_cdresult_tables(msa, from, to, via = NULL, ids = NULL)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{from}{Data table name to start from}

\item{to}{Target data table name}

\item{via}{Character vector of data table name(s) that should be included in
database traversal path, in the order they should be traversed}

\item{ids}{Compound IDs to get matches for, or NULL to get all matches}
}
\value{
A list containing a tibble with ID values relating items within all
  returned tables, plus a list containing each table in the traversal path.
}
\description{
Queries a Compound Discoverer mass spec alignment database to get a series of
related data tables for some or all compounds. If data associated with
specific compound IDs are requested, the database traversal path will start
at the ConsolidatedUnknownCompoundItems table regardless of the specified
starting table, but the additional tables will not be returned.
}
\note{
Compound Discoverer software is produced by Thermo Fisher Scientific. This
package is not affiliated with Thermo Fisher Scientific in any way. For an
official Python interface to Compound Discoverer alignment files, see
\url{https://github.com/thermofisherlsms/pyeds}
}
