% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_alignment_functions.R
\name{get_chromatogram_peaks}
\alias{get_chromatogram_peaks}
\title{Retrieve chromatogram peaks associated with compounds}
\usage{
get_chromatogram_peaks(msa, ids = NULL)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{ids}{Compound IDs to get peaks for, or NULL to get all peaks}
}
\value{
A tibble with the chromatographic peaks
}
\description{
Queries a mass spec alignment database to get all the chromatographic peaks
associated with some or all compounds. This includes all identified adducts.
}
