% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_alignment_functions.R
\name{get_compound_best_ions}
\alias{get_compound_best_ions}
\title{Get best ions associated with compounds}
\usage{
get_compound_best_ions(msa, ids = NULL)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{ids}{Compound IDs to get best ions for, or NULL to get all best ions}
}
\value{
A tibble with the best ion data
}
\description{
Queries a mass spec alignment database to get what the software considers to
be the "best representative compound data" for each compound.
}
\details{
The BestHitType column can be 0 (unknown), 1 (best MS1), 2 (best MS2), or
4 (best deconvoluted MS).
}
