% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_alignment_functions.R
\name{get_compound_ions}
\alias{get_compound_ions}
\title{Get all ions associated with particular compounds}
\usage{
get_compound_ions(msa, ids = NULL)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{ids}{Compound IDs to get ions for, or NULL to get all best ions}
}
\value{
A tibble with the ion data
}
\description{
Queries a mass spec alignment database to get all ions associated with each
compound on a per file basis.
}
