% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_alignment_functions.R
\name{get_compound_spectra}
\alias{get_compound_spectra}
\title{Retrieve spectra associated with compounds}
\usage{
get_compound_spectra(msa, ids = NULL)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{ids}{Compound IDs to get spectra for, or NULL to get all spectra}
}
\value{
A tibble with the spectra
}
\description{
Queries a mass spec alignment database to get the mass spectra associated with
some or all compounds. The actual spectra are stored as XML blobs which can
be extracted using extract_spectral_blob. Note that this method does NOT return
all the spectra in the alignment. Some spectra may not be associated with a
particular compound.
}
