% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compound_xics.R
\name{get_compound_xics}
\alias{get_compound_xics}
\title{Retrieve extracted ion chromatograms (XICs) associated with compounds}
\usage{
get_compound_xics(msa, ids = NULL)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{ids}{Compound IDs to get XICs for, or NULL to get all XICs}
}
\value{
A tibble with the XIC data
}
\description{
Queries a mass spec alignment database to get all the extracted ion chromatograms
associated with some or all compounds. This includes all identified adducts
and isotopologues.
}
