% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_alignment_functions.R
\name{get_data_table_info}
\alias{get_data_table_info}
\title{Get descriptive info for tables and columns in the alignment}
\usage{
get_data_table_info(msa)
}
\arguments{
\item{msa}{An ms_alignment object to query}
}
\value{
A list of two tibbles, one with the table annotation data and the other
 with the categorical data types info
}
\description{
Queries a mass spec alignment database to get annotations for all top-level
data tables and their columns. Also gets level annotations associated with
categorical ("Enum") column types.
}
