% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_alignment_functions.R
\name{get_elemental_compositions}
\alias{get_elemental_compositions}
\title{Retrieve predicted elemental compositions}
\usage{
get_elemental_compositions(msa, ids = NULL)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{ids}{Compound IDs to get compositions for, or NULL to get all predictions}
}
\value{
A tibble with the predicted composition data
}
\description{
Queries a mass spec alignment database to get the predicted elemental
compositions for some or all compounds.
}
