% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_alignment_functions.R
\name{get_file_rt_corrections}
\alias{get_file_rt_corrections}
\title{Get retention time corrections for scans in a mass spec alignment}
\usage{
get_file_rt_corrections(msa, file_id = NULL)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{file_id}{ID numbers for the file retention time corrections to return,
or NULL to return all of them}
}
\value{
A tibble containing the original and corrected retention times for
each file in minutes.
}
\description{
Certain mass spec data alignment formats store retention time correction data
as SQLite blobs containing the data in binary form. This function extracts the
stored data from such blobs. The returned retention times are presumably for
each scan in the raw data file.
}
