% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mzcloud_sim_matches.R
\name{get_mzcloud_sim_matches}
\alias{get_mzcloud_sim_matches}
\title{Retrieve mzCloud library similarity matches}
\usage{
get_mzcloud_sim_matches(msa, ids = NULL)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{ids}{Compound IDs to get matches for, or NULL to get all matches}
}
\value{
A tibble with the mzCloud similarity results, or NA if this alignment
 database does not contain such data. An empty tibble indicates that similarity
 search results exist, but none of the requested compounds have such matches.
}
\description{
Queries a Compound Discoverer mass spec alignment database to get the
mzCloud spectral library similarity search results for some or all compounds.
}
