% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_alignment_functions.R
\name{get_spectral_cloud_matches}
\alias{get_spectral_cloud_matches}
\title{Retrieve cloud spectral library matches}
\usage{
get_spectral_cloud_matches(msa, ids = NULL)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{ids}{Compound IDs to get matches for, or NULL to get all matches}
}
\value{
A tibble with the cloud library matches
}
\description{
Queries a mass spec alignment database to get the cloud-based spectral libary
matches for some or all compounds.
}
