% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupConsecutivePPM.R
\name{groupConsecutivePPM}
\alias{groupConsecutivePPM}
\title{Fast value clustering with PPM threshold}
\usage{
groupConsecutivePPM(x, maxPPM = 5)
}
\arguments{
\item{x}{sorted numeric vector to be grouped}

\item{maxPPM}{threshold below which features are clustered into the same group}
}
\value{
An integer vector of group ID assignments for each value in x.
}
\description{
This is a lightly adapted version of \link[MsFeatures]{groupConsecutive}
from the MsFeatures package. Instead of an absolute difference threshold, it
uses a PPM threshold.
}
