% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrtools.R
\name{hunt}
\alias{hunt}
\title{grab your bows and arrows cuz it's hunting season}
\usage{
hunt(huntdown = NULL)
}
\arguments{
\item{user}{user you want to "hunt down"}
}
\value{
A data table of user and memory usage in GB.
}
\description{
Hunt function to assess local memory usage (mskilab server)
}
\examples{
hunt()
}
