% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ms_alignment.R
\name{import_ms_alignment}
\alias{import_ms_alignment}
\title{Import the data in a mass spec data alignment}
\usage{
import_ms_alignment(f)
}
\arguments{
\item{f}{Path to file containing an alignment project in SQLite format.}
}
\value{
A list containing the alignment data tables.
}
\description{
Certain mass spec data alignment results are stored as SQLite databases. This
function imports relevant result tables for further analysis.
}
