% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzmine_xics.R
\name{import_mzmine_xics}
\alias{import_mzmine_xics}
\title{Import extracted ion chromatograms exported from MZmine}
\usage{
import_mzmine_xics(file)
}
\arguments{
\item{file}{Name of file with XIC data}
}
\value{
A tibble of all the XIC data
}
\description{
This function loads extracted ion chromatogram (XIC) data exported from MZmine
and converts it into a tidy format that works well with the tidyverse.
}
