% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_spectral_library.R
\name{import_spectral_library}
\alias{import_spectral_library}
\title{Import the data in a spectral library}
\usage{
import_spectral_library(f)
}
\arguments{
\item{f}{Path to file containing a spectral library in SQLite format.}
}
\value{
A list containing the library data tables.
}
\description{
Certain spectral libraries are stored as SQLite databases. This function
imports all spectra in a single library and converts the data into a tidy
format.
}
