% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{knn_trim}
\alias{knn_trim}
\title{Trim Weighted Graph to k Nearest Neighbors}
\usage{
knn_trim(graph, k = 3, method = "max", edge_mode = "all")
}
\arguments{
\item{graph}{the igraph object to be trimmed}

\item{k}{the number of neighbors to keep}

\item{method}{"max" to keep k largest edge weights, "min" to keep the k
smallest edge weights}

\item{edge_mode}{"out" to consider only outbound edges, "in" to consider only
inbound edges, or "all" to consider both edge types. Ignored for
un-directed graphs.}
}
\value{
a trimmed copy of the input graph
}
\description{
Given an igraph object with weighted edges, this function will create a
trimmed igraph with only the k nearest neighbors of each vertex. In this
context, nearest neighbors are those with the largest (or smallest)
edge weights.
}
\details{
This is a really simplistic kludge. Use at your own risk!
}
