% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{label_pretty_scientific}
\alias{label_pretty_scientific}
\title{Pretty scientific notation labeling function for use with ggplot2 scales.}
\usage{
label_pretty_scientific(
  digits = 3,
  prefix = "",
  suffix = "",
  decimal.mark = ".",
  trim = TRUE
)
}
\arguments{
\item{digits}{the maximum number of significant digits to show before the exponent}

\item{prefix}{string to add before each label}

\item{suffix}{string to add after each label}

\item{decimal.mark}{character to be used as the decimal point}

\item{trim}{whether values should be trimmed or else right-justified to a common width}
}
\value{
A function which takes a numeric vector and converts it to pretty labels
}
\description{
Converts supplied breaks into labels using plotmath formatting to generate
pretty scientific notation. A prettier version of [scales::label_scientific()].
}
