% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{lollipop_plot}
\alias{lollipop_plot}
\title{Construct lollipop plots}
\usage{
lollipop_plot(labels, values)
}
\arguments{
\item{labels}{the text labels to display with each numerical value}

\item{values}{the numerical values to display}
}
\value{
a ggplot object containing the plot, so you can fine-tune it as
  desired before displaying it
}
\description{
Displays a set of label/value pairs in a lollipop plot.
}
\details{
There isn't really any error-checking here. Put garbage in, you'll
  get garbage out.
}
