% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cd_alignment.R
\name{map_cd_alignment}
\alias{map_cd_alignment}
\title{Map data table relationships in Compound Discoverer alignment results}
\usage{
map_cd_alignment(msa, tables = FALSE)
}
\arguments{
\item{msa}{An ms_alignment object to query}

\item{tables}{Also return the data relationship tables?}
}
\value{
An igraph object containing the relationship between data tables. If
tables is TRUE, returns a list containing the graph, a table describing all
data tables, and a table describing relationships between tables.
}
\description{
Gets the relationships between alignment data tables.
}
\details{
Compound Discoverer alignment results store data across many tables in an
SQLite database. The database also contains a map of how these tables relate
to each other. This function generates an igraph network from the map.
}
\note{
Compound Discoverer software is produced by Thermo Fisher Scientific. This
package is not affiliated with Thermo Fisher Scientific in any way. For an
official Python interface to Compound Discoverer alignment files, see
\url{https://github.com/thermofisherlsms/pyeds}
}
