% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_functions.R
\name{match_ms2_data}
\alias{match_ms2_data}
\title{Search raw MS2 spectra for specific features}
\usage{
match_ms2_data(msn_data, mz, rt, mode, max_ppm_diff = 25, max_rt_diff = 0.1)
}
\arguments{
\item{msn_data}{Spectral data generated by [MSnbase::readMSData()]}

\item{mz}{Vector of m/z values to search for. NOTE this should NOT be neutral
masses!}

\item{rt}{Vector of retention time values to search for}

\item{mode}{Vector of +/-1 indicating the polarity of the m/z values}

\item{max_ppm_diff}{The largest difference allowed between query masses and
MS2 precursor ion masses, in PPM}

\item{max_rt_diff}{The largest difference allowed between query and MS2 scan
retention times, in minutes}
}
\value{
A tibble containing data on all scans in the MS2 files that match any
of the query features.
}
\description{
Thermo RAW files can be converted to mzML format with ProteoWizard and loaded
into R with [MSnbase::readMSData()]. This function searches imported MS2 data
for spectra with particular precursor m/z ratios at specific retention times.
This allows one to quickly find all spectra associated with a given feature
or set of features.
}
