% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ms_features.R
\name{match_ms_features}
\alias{match_ms_features}
\title{Match two sets of mass and retention time pairs}
\usage{
match_ms_features(
  mz1,
  rt1,
  mz2,
  rt2,
  ppm_range = c(-5, 5),
  drt_range = c(0.5, 0.5)
)
}
\arguments{
\item{mz1}{First set of mass or m/z values}

\item{rt1}{First set of retention times}

\item{mz2}{Second set of mass or m/z values}

\item{rt2}{Second set of retention times}

\item{ppm_range}{A length-two numeric giving the minimum and maximum allowed
differences between feature masses in parts per million}

\item{drt_range}{A length-two numeric giving the minimum and maximum allowed
differences between feature retention times}
}
\value{
A tibble of feature matches
}
\description{
Compares two sets of mass/retention time pairs and finds any matches within
a specified range of mass and retention time differences. WARNING: make sure
that both sets are using the same measure of feature mass. This function will
happily compare m/z values to predicted neutral masses and return bogus
results, since there's no way to differentiate them.
}
