% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_functions.R
\name{match_msp}
\alias{match_msp}
\title{Search MSP spectra for specific features}
\usage{
match_msp(msp_data, mz, rt, max_ppm_diff = 25, max_rt_diff = 0.1)
}
\arguments{
\item{msp_data}{Spectral data generated by [read_msp()]}

\item{mz}{Vector of m/z values to search for. NOTE this should NOT be neutral
masses!}

\item{rt}{Vector of retention time values to search for}

\item{max_ppm_diff}{The largest difference allowed between query masses and
MSP masses, in PPM}

\item{max_rt_diff}{The largest difference allowed between query and MSP
retention times, in minutes}
}
\value{
A tibble containing all features in the MSP file that match one of
the query features.
}
\description{
Feature spectra can be exported from Compound Discoverer alignment results by
exporting all features to an mzVault library and exporting that library from
mzVault as an MSP file. This function searches an imported MSP library for
spectra with particular precursor m/z ratios at specific retention times.
This allows one to quickly find all spectra associated with a given feature
or set of features.
}
