% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{misclassification_error}
\alias{misclassification_error}
\title{Calculate Misclassification for a glmnet Model}
\usage{
misclassification_error(mdl, x, y)
}
\arguments{
\item{mdl}{a previously fitted glmnet model}

\item{x}{a vector of values for which to make predictions}

\item{y}{the reference category assignments for the samples in x}
}
\value{
a matrix containing the misclassification error for each lambda value
  in the original model. Matrix has one row per sample and one column per
  lambda.
}
\description{
This function calculates the proportion of a given set of samples that are
mis-classified by a glmnet model. This is particularly useful for calculating
model performance on hold-out sets.
}
