% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicative_imputation.R
\name{multiplicative_imputation}
\alias{multiplicative_imputation}
\title{Impute zeroes in compositional data}
\usage{
multiplicative_imputation(X, imp_factor = 1e-11)
}
\arguments{
\item{X}{non-negative count or relative abundance matrix where rows are
samples and columns are features}

\item{imp_factor}{small proportion to replace zero values with (default 1e-11)}
}
\value{
A matrix with data closed and imputed zeroes
}
\description{
Impute zeroes in compositional data using the multiplicative replacement
strategy described by Martin-Fernandez et al. (2003). Data will be closed
before zero replacement.
}
\references{
Martin-Fernandez, J. A., Barcelo-Vidal, C., & Pawlowsky-Glahn, V. (2003). Dealing
with Zeros and Missing Values in Compositional Data Sets Using Nonparametric
Imputation. Mathematical Geology, 26.
}
