% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chemical_formula.R
\name{parse_chemical_formula}
\alias{parse_chemical_formula}
\title{Parse a chemical formula}
\usage{
parse_chemical_formula(cf)
}
\arguments{
\item{cf}{A character vector with one or more formulas.}
}
\value{
A tibble with one row per formula and one column per element
  containing the total number of each element per formula.
}
\description{
Given a chemical formula string, this function extracts the frequencies of
all elements in the formula. Note that formulas with other text besides
element abbreviations and numbers will likely not parse correctly. Element
abbreviations MUST be properly capitalized.
}
