% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{pick_roc_threshold}
\alias{pick_roc_threshold}
\title{Calculate Optimal ROC Thresholds}
\usage{
pick_roc_threshold(cases, controls)
}
\arguments{
\item{cases}{a vector of values taken by the case samples}

\item{controls}{a vector of values taken by the control samples}
}
\value{
The optimal threshold value.
}
\description{
Calculates the 'best' threshold for a given ROC curve by choosing the
threshold with highest (sensitivity+specificity), using sensitivity to break
ties.
}
