% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot a single mass spectrum}
\usage{
plot_spectrum(spec, label_thresh = NULL)
}
\arguments{
\item{spec}{Spectrum to plot. Must have columns named
'mz' and 'intensity'.}

\item{label_thresh}{Minimum intensity for labeled fragments. Set to NULL for
no labels.}
}
\value{
The ggplot object for rendering or further modification.
}
\description{
Plots one mass spectrum with fragments labeled above a certain threshold
}
