% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_functions.R
\name{plot_spectrum_2d}
\alias{plot_spectrum_2d}
\title{Plot 2D intensity image for MS data}
\usage{
plot_spectrum_2d(ms_data, mz_binwidth = 1)
}
\arguments{
\item{ms_data}{a MSnExp object containing one or more samples}

\item{mz_binwidth}{the m/z bin width (in Da) to use when generating the plot}
}
\value{
list of ggplot objects, one per sample
}
\description{
Given an MSnExp object, generate a 2D plot with retention time on the x-axis
and m/z values on the y-axis. Color indicates signal intensity.
}
