% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plots.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Create a volcano plot}
\usage{
plot_volcano(volcano_data, fc_thresh = NA, p_value_thresh = NA)
}
\arguments{
\item{volcano_data}{Matrix or data frame with at least columns named
'fold_change' and 'p_adj'}

\item{fc_thresh}{The fold-change threshold to display on the plot, or NA for none}

\item{p_value_thresh}{The p-value threshold to display on the plot, or NA for none}
}
\value{
A ggplot object containing the graph
}
\description{
Given a set of fold-changes and p-values, makes and returns a volcano plot.
}
