% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.kmeans.R
\name{predict.kmeans}
\alias{predict.kmeans}
\title{Assign new data to clusters in existing K-means model}
\usage{
\method{predict}{kmeans}(object, newdata, method = c("centers", "classes"), ...)
}
\arguments{
\item{object}{a fitted K-means model object}

\item{newdata}{a matrix-like object with one sample per row and the same
column order as the data used for the original K-means model}

\item{method}{whether to return the centers of the assigned clusters
("centers") or the assigned cluster IDs ("classes")}

\item{...}{not used}
}
\value{
An array containing either predicted cluster centers or cluster IDs
}
\description{
Assigns new samples to the clusters defined by an existing K-means model.
}
