% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mgf.R
\name{read_mgf}
\alias{read_mgf}
\title{Read MGF files from the GNPS spectral library}
\usage{
read_mgf(path)
}
\arguments{
\item{path}{The location of the MGF file to import}
}
\value{
A list of entries where each entry has the header information as a
  list of key-value pairs and the spectrum itself as a matrix with m/z and
  intensity columns.
}
\description{
Parses the flavor of MGF file generated by the GNPS
spectral library.
}
