% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mona_msp.R
\name{read_mona_msp}
\alias{read_mona_msp}
\title{Read MSP files from MoNA spectral library}
\usage{
read_mona_msp(path)
}
\arguments{
\item{path}{The location of the MSP file to import}
}
\value{
A list of entries where each entry has the header information as a
  list of key-value pairs and the spectrum itself as a matrix with m/z and
  intensity columns.
}
\description{
Parses the flavor of MSP file generated by the MassBank of North America
(MoNA) spectral library. This parser also handles MSP files generated by the
GNPS spectral library.
}
