% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_functions.R
\name{read_msp}
\alias{read_msp}
\title{Read MSP file of feature spectra}
\usage{
read_msp(path)
}
\arguments{
\item{path}{The location of the MSP file to import}
}
\value{
A list containing precursor m/z values, retention times, scan indices,
scan modes, raw header data, and spectra for each entry in the MSP file.
}
\description{
Feature spectra can be exported from Compound Discoverer alignment results by
exporting all features to an mzVault library and exporting that library from
mzVault as an MSP file. This function reads the spectral data into a useful
format.
}
