% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.kmeans.R
\name{reorder.kmeans}
\alias{reorder.kmeans}
\title{Change the order of clusters in existing K-means model}
\usage{
\method{reorder}{kmeans}(x, order, ...)
}
\arguments{
\item{x}{a fitted K-means model object}

\item{order}{a vector of the same length as the number of clusters, giving
the new order of the current cluster IDs}

\item{...}{not used}
}
\value{
A reordered copy of the K-means model object
}
\description{
Changes the order (and associated IDs) of the clusters defined by an
existing K-means model.
}
