% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_functions.R
\name{save_target_feature_table}
\alias{save_target_feature_table}
\title{Export features for import to ID-X method editor}
\usage{
save_target_feature_table(
  raw_data,
  file,
  rt_range = 0.5,
  mode = c("both", "positive", "negative")
)
}
\arguments{
\item{raw_data}{a data.frame or equivalent formatted like Compound Discoverer
export files}

\item{file}{Base file name to write to. Ionization polarity and file
extension will be added automatically.}

\item{rt_range}{Half width of the retention time window around the feature's
observed retention time, in minutes.}

\item{mode}{Only export features with the given polarity ("positive",
"negative", or "both"). Default is "both".}
}
\value{
Invisibly returns a list with the formatted table(s), separated by
 polarity.
}
\description{
The method editor software for the ID-X can import properly-formatted TSV
files into the Targeted Mass In/Exclusion nodes. This function accepts rows
from the raw Compound Discoverer output format and saves each feature in the
TSV format.
}
\section{Note}{


Versions of Compound Discoverer <3.3 don't include raw m/z values or polarity
in the exported data, so this method will not work for older datasets.
}

