% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_adduct_candidates.R
\name{score_adduct_candidates}
\alias{score_adduct_candidates}
\title{Score all possible adduct identities for an MS1 peak}
\usage{
score_adduct_candidates(
  mz,
  mz_polarity,
  ms1_spec,
  all_adducts,
  ppm_thresh = 15
)
}
\arguments{
\item{mz}{The m/z ratio of interest}

\item{mz_polarity}{The polarity of the m/z ratio. Must be either "positive" or
"negative"}

\item{ms1_spec}{The MS1 spectrum being examined. Must be tibble-alike with
columns "polarity", "mz", and "intensity"}

\item{all_adducts}{The possible adducts to be considered. Must be tibble-alike
with columns "polarity", "Ion name", "Power", "Mult", and "Mass".}

\item{ppm_thresh}{The maximum difference between observed m/z values and
proposed adduct m/z values}
}
\value{
A tibble of adduct candidate match scores.
}
\description{
Given a peak from an MS1 spectrum, tests how well different adduct assignments
match the MS1 spectrum.
}
