% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_functions.R
\name{tidy_gap_fill_methods}
\alias{tidy_gap_fill_methods}
\title{Tidy Compound Discoverer Gap Filling Information}
\usage{
tidy_gap_fill_methods(method_data_file)
}
\arguments{
\item{method_data_file}{the path to a CSV file containing the raw data}
}
\value{
the data in a tidy tibble
}
\description{
Compound Discoverer exports detailed gap filling information in a
nearly-useless format. This method converts it into a tidy format for use
with the other data exported from Compound Discoverer.
}
\details{
Compound Discoverer abuses Excel's row grouping feature to export
  information on how gaps were filled after peak detection. This method
  cleans it up into a format R can use. You MUST export ONLY the Name,
  Molecular Weight, and RT columns from the Compounds table and ONLY the Fill
  Status and Study File ID columns from the Filled Gaps table. Then open the
  XLSX file in Excel and save it as a CSV file.
}
