% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_functions.R
\name{tidy_mzCloud_hits}
\alias{tidy_mzCloud_hits}
\title{Tidy Compound Discoverer mzCloud Hits}
\usage{
tidy_mzCloud_hits(mzcloud_data)
}
\arguments{
\item{mzcloud_data}{a tibble containing the exported mzCloud data}
}
\value{
the data in a tidy tibble with data nesting removed
}
\description{
Compound Discoverer exports detailed information on mzCloud database matches
into a weird nested Excel format that doesn't work well with R. This method
converts it into a tidy format for use with the other data exported from
Compound Discoverer.
}
\details{
Compound Discoverer abuses Excel's row grouping feature to export
  information on mzCloud database hits. This method cleans it up into a
  format R can use.
}
