% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spectral_blob.R
\name{extract_spectral_blob}
\alias{extract_spectral_blob}
\title{Extract spectral data from an SQLite blob}
\usage{
extract_spectral_blob(blb, zip_dir = tempdir())
}
\arguments{
\item{blb}{A raw vector containing an SQLite blob of spectral data}

\item{zip_dir}{Directory where XML files should be unzipped}
}
\value{
A list containing the library data tables.
}
\description{
Certain mass spec data alignment formats store feature spectra as SQLite
blobs containing a compressed XML file. This function extracts the stored
data from such blobs.
}
\details{
The XML file contains a wealth of metadata about the spectrum, which is not
currently extracted by this function.
}
